Refactor code and create completion endpoint
