+++ freelance4u/src/main/java/ch/zhaw/freelance4u/service/JobService.java
@@ -12,7 +12,7 @@ import ch.zhaw.freelance4u.repository.JobRepository;
 @Service
 public class JobService {
     @Autowired
-    JobRepository jobRepository;
+    private JobRepository jobRepository;
 
     public Optional<Job> assignJob(String jobId, String freelancerId) {
         Job job = this.jobRepository.findById(jobId).orElse(null);
@@ -30,4 +30,12 @@ public class JobService {
         return Optional.of(this.jobRepository.save(job));
     }
 
+    public Optional<Job> completeJob(String jobId, String freelancerId) {
+        return this.jobRepository.findById(jobId)
+                .filter(job -> job.getJobState() == JobState.ASSIGNED && job.getFreelancerId().equals(freelancerId))
+                .map(job -> {
+                    job.setJobState(JobState.DONE);
+                    return this.jobRepository.save(job);
+                });
+    }
 }
